Product page adding swiper on photos and fixing back button (cosmetic)
