+++ src/components/Product/ProductPhotos/ProductPhotos.tsx
@@ -1,13 +1,10 @@
-import React, { useEffect, useRef, useState } from 'react';
+import React, { useEffect, useState } from 'react';
 import cn from 'classnames';
-import { Swiper, SwiperClass, SwiperSlide } from 'swiper/react';
-import { Navigation } from 'swiper/modules';
 
 import styles from './ProductPhotos.module.scss';
+import ProductPhotoSwiper from './ProductPhotoSwiper';
 
 import { Product } from '@/types/Product.type';
-
-import 'swiper/css/navigation';
 
 type Props = {
   product: Product;
@@ -15,13 +12,6 @@ type Props = {
 
 const ProductPhotos: React.FC<Props> = ({ product }) => {
   const [selectedImage, setSelectedImage] = useState(product?.images[0]);
-  const swiperRef = useRef<SwiperClass | null>(null);
-
-  const handleSlideChange = (swiper: SwiperClass) => {
-    const currentIndex = swiper.activeIndex;
-
-    setSelectedImage(product.images[currentIndex]);
-  };
 
   useEffect(() => {
     setSelectedImage(product?.images[0]);
@@ -30,30 +20,11 @@ const ProductPhotos: React.FC<Props> = ({ product }) => {
   return (
     <>
       <div className={styles.mainImageBox}>
-        <Swiper
-          modules={[Navigation]}
-          spaceBetween={100}
-          slidesPerView={1}
-          navigation={{
-            nextEl: `.${styles.nextButton}`,
-            prevEl: `.${styles.prevButton}`,
-          }}
-          onSwiper={(swiper) => {
-            swiperRef.current = swiper;
-          }}
-          onSlideChange={handleSlideChange}
-          pagination={false}
-        >
-          {product.images.map((_, index) => (
-            <SwiperSlide key={index}>
-              <img
-                src={`/${selectedImage}`}
-                alt="product image"
-                className={styles.mainPhoto}
-              />
-            </SwiperSlide>
-          ))}
-        </Swiper>
+        <ProductPhotoSwiper
+          product={product}
+          selectedImage={selectedImage}
+          setSelectedImage={setSelectedImage}
+        />
       </div>
 
       <div className={styles.imagesContainer}>
